+++ src/features/suggestions/search/foodNameSearch.ts
@@ -3,7 +3,7 @@ import { Serving } from "../../../model/Food";
 import portions from "../portion/portions";
 import { buildDocuments, search, autoSuggest } from "./foodNameMiniSearch";
 import servings from "../serving/servings";
-import parseAmount from "../parser/amount";
+import parseAmount from "../parser/DecomposedAmount";
 import { isMeasurementConvertible, Unit } from "../convert";
 
 export interface PredefinedSuggestion {
